+++ src/onegov.town/onegov/town/models/__init__.py
@@ -6,6 +6,7 @@ from onegov.town.models.form import BuiltinFormDefinition, CustomFormDefinition
 from onegov.town.models.image import Image, ImageCollection, Thumbnail
 from onegov.town.models.page import AtoZPages, News, Topic
 from onegov.town.models.page_move import PageMove
+from onegov.town.models.person_move import PersonMove
 from onegov.town.models.resource import DaypassResource
 from onegov.town.models.search import Search
 from onegov.town.models.sitecollection import SiteCollection
@@ -27,6 +28,7 @@ __all__ = [
     'ImageCollection',
     'News',
     'PageMove',
+    'PersonMove',
     'Search',
     'SiteCollection',
     'Thumbnail',