+++ src/onegov.town/onegov/town/models/extensions.py
@@ -135,9 +135,13 @@ class ContactExtension(ContentExtension):
     @contact.setter
     def contact(self, value):
         self.content['contact'] = value
-        self.content['contact_html'] = '<ul><li>{}</li></ul>'.format(
-            '</li><li>'.join(linkify(value).splitlines())
-        )
+
+        if value:
+            self.content['contact_html'] = '<ul><li>{}</li></ul>'.format(
+                '</li><li>'.join(linkify(value).splitlines())
+            )
+        else:
+            self.content['contact_html'] = ""
 
     @property
     def contact_html(self):
@@ -174,52 +178,163 @@ class PersonLinkExtension(ContentExtension):
 
     @property
     def people(self):
-        if 'people' in self.content:
-            people = OrderedDict(self.content['people'])
+        """ Returns the people linked to this content or None.
 
-            if len(people):
-                query = PersonCollection(object_session(self)).query()
-                query = query.filter(Person.id.in_(people.keys()))
-                query = query.order_by(Person.last_name, Person.first_name)
+        The context specific function is temporarily stored on the
+        ``context_specific_function`` attribute on each object in the
+        resulting list.
 
-                result = query.all()
+        """
 
-                for person in result:
-                    person.context_specific_function = people[person.id.hex]
+        if not self.content.get('people'):
+            return None
 
-                return result
+        people = OrderedDict(self.content['people'])
+
+        query = PersonCollection(object_session(self)).query()
+        query = query.filter(Person.id.in_(people.keys()))
+
+        result = []
+
+        for person in query.all():
+            person.context_specific_function = people[person.id.hex]
+            result.append(person)
+
+        order = list(people.keys())
+        result.sort(key=lambda p: order.index(p.id.hex))
+
+        return result
 
     def get_selectable_people(self, request):
+        """ Returns a list of people which may be linked. """
+
         query = PersonCollection(request.app.session()).query()
         query = query.order_by(Person.last_name, Person.first_name)
 
         return query.all()
 
+    def get_person_function_by_id(self, id):
+        for _id, function in self.content.get('people', []):
+            if id == _id:
+                return function
+
+    def move_person(self, subject, target, direction):
+        """ Moves the subject below or above the target.
+
+        :subject:
+            The key of the person to be moved.
+
+        :target:
+            The key of the person above or below which the subject is moved.
+
+        :direction:
+            The direction relative to the target. Either 'above' or 'below'.
+
+        """
+        assert direction in ('above', 'below')
+        assert subject != target
+        assert self.content.get('people')
+
+        def new_order():
+            subject_function = self.get_person_function_by_id(subject)
+            target_function = self.get_person_function_by_id(target)
+
+            for person, function in self.content['people']:
+
+                if person == subject:
+                    continue
+
+                if person == target and direction == 'above':
+                    yield subject, subject_function
+                    yield target, target_function
+                    continue
+
+                if person == target and direction == 'below':
+                    yield target, target_function
+                    yield subject, subject_function
+                    continue
+
+                yield person, function
+
+        self.content['people'] = list(new_order())
+
     def extend_form(self, form_class, request):
+
+        # XXX this is kind of implicitly set by the builder
+        fieldset_id = 'people'
+        fieldset_label = _("People")
 
         class PeoplePageForm(form_class):
 
             def get_people_fields(self, with_function):
                 for field_id, field in self._fields.items():
-
-                    # XXX this is kind of implicitly set by the fieldset
-                    if field_id.startswith('people_'):
+                    if field_id.startswith(fieldset_id):
                         if with_function or not field_id.endswith('_function'):
                             yield field_id, field
 
-            def get_people_and_function(self):
+            def get_people_and_function(self, selected_only=True):
                 fields = self.get_people_fields(with_function=False)
 
                 for field_id, field in fields:
-                    if field.data is True:
+                    if not selected_only or field.data is True:
                         person_id = field.id.hex
                         function = self._fields[field_id + '_function'].data
 
                         yield person_id, function
 
+            def is_ordered_people(self, existing_people):
+                """ Returns True if the current list of people is ordered
+                from A to Z.
+
+                """
+                if not existing_people:
+                    return True
+
+                existing_people = OrderedDict(existing_people)
+                ordered_people = OrderedDict(self.get_people_and_function(
+                    selected_only=False
+                ))
+
+                sorted_existing_people = sorted(
+                    [key for key in existing_people],
+                    key=list(ordered_people.keys()).index
+                )
+
+                return list(existing_people.keys()) == sorted_existing_people
+
             def update_model(self, model):
+                previous_people = model.content.get('people', [])
+
                 super().update_model(model)
-                model.content['people'] = list(self.get_people_and_function())
+
+                if self.is_ordered_people(previous_people):
+                    # if the people are ordered a-z, we take the ordering from
+                    # get_people_and_function, which comes by A-Z already
+                    model.content['people'] = list(
+                        self.get_people_and_function()
+                    )
+                else:
+                    # if the people are not ordered we keep the order of the
+                    # existing list and add the new people at the end
+                    existing = set()
+                    selected = {
+                        key for key, function
+                        in self.get_people_and_function()
+                    }
+
+                    old_people = list()
+                    new_people = list()
+
+                    for id, function in previous_people:
+                        if id in selected:
+                            old_people.append((id, function))
+                            existing.add(id)
+
+                    for id, function in self.get_people_and_function():
+                        if id not in existing:
+                            new_people.append((id, function))
+
+                    model.content['people'] = old_people + new_people
 
             def apply_model(self, model):
                 super().apply_model(model)
@@ -234,7 +349,7 @@ class PersonLinkExtension(ContentExtension):
                             = people[field.id.hex]
 
         builder = WTFormsClassBuilder(PeoplePageForm)
-        builder.set_current_fieldset(_("People"))
+        builder.set_current_fieldset(fieldset_label)
 
         for person in self.get_selectable_people(request):
             field_id = builder.add_field(